Move leaderboard.go into db package
